verilog: Started hacking at the lexer, once again
I think, we might need to ditch using chumsky, as nice as it is, using it for lexing it rough.

The option to use something like `logos` is there, but I think we need to re-think about the multi-standard/multi-variant design, not sure if we're going about this the right way.

One option is to just emit a token stream that should be valid for everything up to the newest standard for all variants, and then deal with validation on the parse side of things, I'm not sure.